requirements.txt
ccxt==4.4.12
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.4
pyyaml==6.0.2
matplotlib==3.9.0
ta==0.11.0
